+++ packages/core/src/components.d.ts
@@ -80,6 +80,20 @@ export namespace Components {
          */
         "value": string;
     }
+    interface DsoDropdownButton {
+        /**
+          * The text of the dropdown button.
+         */
+        "label": string;
+        /**
+          * Whether the dropdown is open or closed. This attribute is reflected and mutable.
+         */
+        "open": boolean;
+        /**
+          * The button style variant
+         */
+        "variant": "primary" | "secondary" | "tertiary";
+    }
     interface DsoHighlightBox {
         "border"?: boolean;
         "dropShadow"?: boolean;
@@ -199,6 +213,12 @@ declare global {
         prototype: HTMLDsoDatePickerElement;
         new (): HTMLDsoDatePickerElement;
     };
+    interface HTMLDsoDropdownButtonElement extends Components.DsoDropdownButton, HTMLStencilElement {
+    }
+    var HTMLDsoDropdownButtonElement: {
+        prototype: HTMLDsoDropdownButtonElement;
+        new (): HTMLDsoDropdownButtonElement;
+    };
     interface HTMLDsoHighlightBoxElement extends Components.DsoHighlightBox, HTMLStencilElement {
     }
     var HTMLDsoHighlightBoxElement: {
@@ -277,6 +297,7 @@ declare global {
         "dso-badge": HTMLDsoBadgeElement;
         "dso-banner": HTMLDsoBannerElement;
         "dso-date-picker": HTMLDsoDatePickerElement;
+        "dso-dropdown-button": HTMLDsoDropdownButtonElement;
         "dso-highlight-box": HTMLDsoHighlightBoxElement;
         "dso-icon": HTMLDsoIconElement;
         "dso-info": HTMLDsoInfoElement;
@@ -360,6 +381,20 @@ declare namespace LocalJSX {
           * Date value. Must be in Dutch date format: DD-MM-YYYY.
          */
         "value"?: string;
+    }
+    interface DsoDropdownButton {
+        /**
+          * The text of the dropdown button.
+         */
+        "label"?: string;
+        /**
+          * Whether the dropdown is open or closed. This attribute is reflected and mutable.
+         */
+        "open"?: boolean;
+        /**
+          * The button style variant
+         */
+        "variant"?: "primary" | "secondary" | "tertiary";
     }
     interface DsoHighlightBox {
         "border"?: boolean;
@@ -454,6 +489,7 @@ declare namespace LocalJSX {
         "dso-badge": DsoBadge;
         "dso-banner": DsoBanner;
         "dso-date-picker": DsoDatePicker;
+        "dso-dropdown-button": DsoDropdownButton;
         "dso-highlight-box": DsoHighlightBox;
         "dso-icon": DsoIcon;
         "dso-info": DsoInfo;
@@ -477,6 +513,7 @@ declare module "@stencil/core" {
             "dso-badge": LocalJSX.DsoBadge & JSXBase.HTMLAttributes<HTMLDsoBadgeElement>;
             "dso-banner": LocalJSX.DsoBanner & JSXBase.HTMLAttributes<HTMLDsoBannerElement>;
             "dso-date-picker": LocalJSX.DsoDatePicker & JSXBase.HTMLAttributes<HTMLDsoDatePickerElement>;
+            "dso-dropdown-button": LocalJSX.DsoDropdownButton & JSXBase.HTMLAttributes<HTMLDsoDropdownButtonElement>;
             "dso-highlight-box": LocalJSX.DsoHighlightBox & JSXBase.HTMLAttributes<HTMLDsoHighlightBoxElement>;
             "dso-icon": LocalJSX.DsoIcon & JSXBase.HTMLAttributes<HTMLDsoIconElement>;
             "dso-info": LocalJSX.DsoInfo & JSXBase.HTMLAttributes<HTMLDsoInfoElement>;