+++ packages/sources/src/components/dropdown-button/dropdown-button.models.ts
@@ -1,0 +1,4 @@
+export interface DropdownButton {
+  label: string;
+  variant: "primary" | "secondary" | "tertiary";
+}